requirements.txt
aenum==2.2.4
bandit==1.7.1
boto3==1.21.6
botocore==1.24.6
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cryptography==37.0.0
defusedxml~=0.7.1
docutils==0.14
fabric~=2.5.0
filelock==3.0.12
flask==1.1.2
gevent~=21.1.2
greenlet==0.4.17
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.3
jira~=3.0.1
jmespath==0.9.4
jproperties==2.1.0
jsonschema==3.2.0
kubernetes==21.7.0
locust==1.4.3
MarkupSafe==1.1.1
mdstat==1.0.4
msgpack==0.6.2
munch==2.5.0
mypy==0.782
numpy==1.21.5
pandas==1.2.1
paramiko==2.10.4
prometheus-client
pika==1.1.0
pycryptodome==3.10.1
PyHamcrest==2.0.2
pylint==2.12.2
pymongo~=3.11.4
pysftp==0.2.9
pytest~=6.2.1
pytest-cache==1.0
pytest-cov==2.11.1
pytest-forked==1.3.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-ordering==0.6
pytest-parallel==0.1.0
pytest-xdist==2.3.0
python-dateutil==2.8.0
python-jenkins~=1.7.0
pyyaml==6.0.0
pyzmq==18.0.2
pyOpenSSL==22.0.0
requests>=2.22.0,<2.27.1
schedule
setuptools==41.6.0
strip-ansi==0.1.1
testfixtures~=6.17.0
urllib3==1.26.5
Werkzeug==1.0.1
word2number~=1.1
xmltodict==0.12.0
asyncio==3.4.3
cerberus
assertpy